requirements.txt
diffusers==0.32.1
transformers==4.46.3
torchvision==0.19.0
torchmetrics==1.6.0
einops==0.8.0
matplotlib
